+++ src/main/java/com/halil/bookwebsite/services/impl/ReviewServiceImpl.java
@@ -1,6 +1,8 @@
 package com.halil.bookwebsite.services.impl;
 
 import com.halil.bookwebsite.entities.Review;
+import com.halil.bookwebsite.exceptions.AlreadyExistsException;
+import com.halil.bookwebsite.exceptions.NotFoundException;
 import com.halil.bookwebsite.repositories.ReviewRepository;
 import com.halil.bookwebsite.services.ReviewService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,11 +20,17 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public Review update(Review review) {
+        if (!reviewRepository.existsById(review.getId())) {
+            throw new NotFoundException("Review not found");
+        }
         return reviewRepository.save(review);
     }
 
     @Override
     public void deleteByID(Long id) {
+        if (!reviewRepository.existsById(id)) {
+            throw new NotFoundException("Review not found");
+        }
         reviewRepository.deleteById(id);
     }
 }
